Add files via upload
The following implementation code is based on experiencor project (https://github.com/experiencor/keras-yolo3) which contains useful functions such as make_yolov3_model() that will help us to create the model, convblock() to helps us create blocks of layers. The code also contains functions to load the weights of the saved model. This implementation follows the steps given in this tutoarial (https://machinelearningmastery.com/how-to-perform-object-detection-with-yolov3-in-keras/) To use this code simply run this notebook and loacate the place which ask for image example and provide the path to the image your want to use for object detection. You will see the detection results at the end. Please check if the following libraries are imported correctly.